3.14h fixed4 - formule

+++ Aplikace_1_0/Source/ewitis/gui/dfTableRuns.py
@@ -112,8 +112,8 @@ class DfTableRuns(DfTable):
             
     # DELETE BUTTON          
     def sDeleteAll(self):
-        DfTable.sDeleteAll(self)
-        tableTimes.deleteAll() 
+        if DfTable.sDeleteAll(self) == True:
+            tableTimes.deleteAll()
